test: refac data savings tests to options
